requirements.txt
PyMuPDF
pdf2image
Pillow

# Web UI
flask
streamlit>=1.28.0

# Testing
pytest>=7.0.0
